[Library -> Win32] Move P/Invoke to lib.
