+++ GyroShell.Library/Commands/StartFlyoutCommands.cs
@@ -13,12 +13,12 @@ namespace GyroShell.Library.Commands
     public partial class StartFlyoutCommands
     {
         private readonly IEnvironmentInfoService m_envService;
-        private readonly IInternalLauncher m_internalLauncher;
+        private readonly IpublicLauncher m_publicLauncher;
 
-        public StartFlyoutCommands(IEnvironmentInfoService envService, IInternalLauncher internalLauncher)
+        public StartFlyoutCommands(IEnvironmentInfoService envService, IpublicLauncher publicLauncher)
         {
             m_envService = envService;
-            m_internalLauncher = internalLauncher;
+            m_publicLauncher = publicLauncher;
         }
 
         [RelayCommand]
@@ -28,20 +28,20 @@ namespace GyroShell.Library.Commands
 
             if (m_envService.SettingsInstances <= 1)
             {
-                m_internalLauncher.LaunchShellSettings();
+                m_publicLauncher.LaunchShellSettings();
             }
         }
 
         [RelayCommand]
         public void RestartGyroShell()
         {
-            m_internalLauncher.LaunchNewShellInstance();
+            m_publicLauncher.LaunchNewShellInstance();
         }
 
         [RelayCommand]
         public void LaunchTaskManager()
         {
-            m_internalLauncher.LaunchProcess("taskmgr.exe", false, true);
+            m_publicLauncher.LaunchProcess("taskmgr.exe", false, true);
         }
 
         [RelayCommand]
@@ -60,6 +60,36 @@ namespace GyroShell.Library.Commands
         public void LaunchRunDialog()
         {
             Process.Start("explorer.exe", "shell:::{2559a1f3-21d7-11d4-bdaf-00c04f60b9f0}");
+        }
+
+        [RelayCommand]
+        public void SignOutWindows()
+        {
+            // TODO: Implement
+        }
+
+        [RelayCommand]
+        public void SleepWindows()
+        {
+            // TODO: Implement
+        }
+
+        [RelayCommand]
+        public void ShutdownWindows()
+        {
+            m_publicLauncher.LaunchProcess("shutdown /s /t 00", false, true);
+        }
+
+        [RelayCommand]
+        public void RestartWindows()
+        {
+            m_publicLauncher.LaunchProcess("shutdown /r /t 00", false, true);
+        }
+
+        [RelayCommand]
+        public void ShowDesktop()
+        {
+            // TODO: Implement
         }
     }
 }
