+++ GyroShell.Library/Models/InternalData/ModuleModel.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
 
-namespace GyroShell.Library.Models.InternalData
+namespace GyroShell.Library.Models.publicData
 {
     public class ModuleModel : INotifyPropertyChanged
     {
